+++ hari_client/models/models.py
@@ -947,6 +947,10 @@ class MediaCreate(BaseModel):
     back_reference: str
     media_url: str | None = None
 
+    id: str | None = pydantic.Field(
+        default=None, title="Id"
+    )  # used to identify already uploaded medias
+
     archived: bool = False
     scene_id: str | None = None
     realWorldObject_id: str | None = None
@@ -981,6 +985,10 @@ class MediaObjectCreate(BaseModel):
     media_id: str
     source: DataSource = DataSource.REFERENCE
     back_reference: str
+
+    id: str | None = pydantic.Field(
+        default=None, title="Id"
+    )  # used to identify already uploaded medias
 
     archived: bool = False
     scene_id: str | None = None
@@ -1018,7 +1026,10 @@ class BulkMediaObjectCreate(MediaObjectCreate):
 
 
 class AttributeCreate(BaseModel):
-    id: uuid.UUID
+    # This value is mandatory during upload but needs to be synced with existing values in HARI and this is set programmatically in the Uploader
+    id: uuid.UUID | str | None = (
+        None  # value should be UUID but legacy attributes can be str
+    )
     name: str
     annotatable_id: str
     annotatable_type: typing.Literal[