+++ hari_client/utils/upload.py
@@ -2,9 +2,9 @@ import time
 import uuid
 from typing import Tuple
 
-from hari_client import hari_uploader
 from hari_client import HARIClient
 from hari_client import models
+from hari_client import state_aware_hari_uploader as hari_uploader
 from hari_client.models.models import Media
 
 
@@ -64,6 +64,8 @@ def check_and_create_dataset(
     """
     datasets = hari.get_datasets()
     dataset_names = [dataset.name for dataset in datasets]
+
+    assert user_group is not None, "User group is required."
 
     if dataset_name not in dataset_names:
         new_dataset = hari.create_dataset(
@@ -138,44 +140,58 @@ def check_and_upload_dataset(
         client=hari, dataset_id=dataset_id, object_categories=object_categories
     )
 
-    uploaded_medias = hari.get_medias(dataset_id)
-    uploaded_back_references = [m.back_reference for m in uploaded_medias]
-
-    elements_to_upload = 0
     for media in medias:
-        if media.back_reference not in uploaded_back_references:
-            uploader.add_media(media)
-            elements_to_upload += 1
+        uploader.add_media(media)
 
-    if elements_to_upload > 0:
-        upload_results = uploader.upload()
+    upload_results = uploader.upload()
 
-        # Inspect upload results
-        print(f"media upload status: {upload_results.medias.status.value}")
-        print(f"media upload summary\n  {upload_results.medias.summary}")
+    # Inspect upload results
+    num_conflict_media = len(
+        [
+            r.status
+            for r in upload_results.medias.results
+            if r.status == models.ResponseStatesEnum.CONFLICT
+        ]
+    )
+    num_conflict_media_objects = len(
+        [
+            r.status
+            for r in upload_results.media_objects.results
+            if r.status == models.ResponseStatesEnum.CONFLICT
+        ]
+    )
+    num_conflict_attributes = len(
+        [
+            r.status
+            for r in upload_results.attributes.results
+            if r.status == models.ResponseStatesEnum.CONFLICT
+        ]
+    )
+    print(f"media upload status: {upload_results.medias.status.value}")
+    print(
+        f"media upload summary\n  {upload_results.medias.summary} due to skipped {num_conflict_media}"
+    )
 
-        print(
-            f"media object upload status: {upload_results.media_objects.status.value}"
-        )
-        print(f"media object upload summary\n  {upload_results.media_objects.summary}")
+    print(f"media object upload status: {upload_results.media_objects.status.value}")
+    print(
+        f"media object upload summary\n  {upload_results.media_objects.summary} due to skipped {num_conflict_media_objects}"
+    )
 
-        print(f"attribute upload status: {upload_results.attributes.status.value}")
-        print(f"attribute upload summary\n  {upload_results.attributes.summary}")
+    print(f"attribute upload status: {upload_results.attributes.status.value}")
+    print(
+        f"attribute upload summary\n  {upload_results.attributes.summary} due to skipped {num_conflict_attributes}"
+    )
 
-        if (
-            upload_results.medias.status != models.BulkOperationStatusEnum.SUCCESS
-            or upload_results.media_objects.status
-            != models.BulkOperationStatusEnum.SUCCESS
-            or upload_results.attributes.status
-            != models.BulkOperationStatusEnum.SUCCESS
-        ):
-            print(
-                "The data upload wasn't fully successful. Subset and metadata creation are skipped. See the details below."
-            )
-            print(f"media upload details: {upload_results.medias.results}")
-
-    else:
-        print("WARNING: No data for upload specified which is not already uploaded.")
+    # check if any failed (not due to skipping/conflict)
+    if (
+        upload_results.medias.summary.failed - num_conflict_media > 0
+        or upload_results.media_objects.summary.failed - num_conflict_media_objects > 0
+        or upload_results.attributes.summary.failed - num_conflict_attributes > 0
+    ):
+        print("The data upload wasn't fully successful. See the details below.")
+        print(f"media upload details: {upload_results.medias.results}")
+        print(f"media objects upload details: {upload_results.media_objects.results}")
+        print(f"attributes upload details: {upload_results.attributes.results}")
 
     new_subset_id, reused = check_and_create_subset_for_all(
         hari, dataset_id, new_subset_name, subset_type